refactoring and pypi urls
